Put plotting logic in separate file + tune selection behaviour
